Clean up analyze module
